django/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.1
django-cas-ng==4.1.1
djangorestframework==3.11.1
idna==2.10
lxml==4.5.2
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-cas==1.5.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
django-cors-headers
